move CMSServerExtension to test framework
