+++ integration-tests/src/test/java/com/condation/cms/e2e/HTMLPipelineTest.java
@@ -20,6 +20,7 @@ package com.condation.cms.e2e;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
+import com.condation.cms.test.e2e.CMSServerExtension;
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.junit.UsePlaywright;
 import org.assertj.core.api.Assertions;
